# transaction-ms/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "stori API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/stori/v1/public",
    "paths": {
        "/account/balance": {
            "post": {
                "description": "upload one csv file with the resume of account user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "account balance processor",
                "parameters": [
                    {
                        "type": "string",
                        "format": "email",
                        "description": "name search by q",
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "type": "file",
                        "description": "this is a csv test file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/infrastructure.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/infrastructure.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "infrastructure.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {},
                "msg": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}